requirements.txt
Django>=6.0,<7.0
djangorestframework>=3.16,<4.0
PyMySQL>=1.1.0
